+++ FSItem-Utilities.m
@@ -18,7 +18,7 @@
 
 @implementation FSItem(Utilities)
 
-- (unsigned) deepFileCountIncludingPackages: (BOOL) lookInPackages
+- (NSUInteger) deepFileCountIncludingPackages: (BOOL) lookInPackages
 {
 	if ( [self isSpecialItem] )
 		return 0;
@@ -52,7 +52,7 @@
 		return nil; //path defines no child
 	
 	//get relative child path (relative to self)
-	unsigned childPathStartIndex = [myPath length];
+	NSUInteger childPathStartIndex = [myPath length];
 	if ( ![myPath isEqualToString: NSOpenStepRootDirectory()] ) // should just be @"/"
 		childPathStartIndex++;
