+++ FSItem.m
@@ -314,7 +314,7 @@ static void InsertChildKeepingSizeOrder( NSMutableArray *children, FSItem *newCh
 		return nil;
 }
 
-- (FSItem*) childAtIndex: (unsigned) index
+- (FSItem*) childAtIndex: (NSUInteger) index
 {
 	if ( ![self isSpecialItem] )
 		return [_childs objectAtIndex: index];
@@ -322,7 +322,7 @@ static void InsertChildKeepingSizeOrder( NSMutableArray *children, FSItem *newCh
 		return nil;
 }
 
-- (unsigned) childCount
+- (NSUInteger) childCount
 {
 	if ( ![self isSpecialItem] )
 		return [_childs count];
@@ -422,7 +422,7 @@ static void InsertChildKeepingSizeOrder( NSMutableArray *children, FSItem *newCh
 		case FileFolderItem:
 			if ( [self isFolder] )
 			{
-				unsigned i = [_childs count];
+				NSUInteger i = [_childs count];
 				while ( i-- )
 				{
 					FSItem *child = [_childs objectAtIndex: i];
@@ -593,7 +593,7 @@ static void InsertChildKeepingSizeOrder( NSMutableArray *children, FSItem *newCh
     //let our childs do the same
 	if ( includingChildren && [self isFolder] )
 	{
-		unsigned i = [self childCount];
+		NSUInteger i = [self childCount];
 		while ( i-- )
 			[[self childAtIndex: i] setKindStringIncludingChildren: YES];
 	}
